requirements.txt
celery
ipdb
ipython
requests-html
python-dotenv
slackclient
redis